Move requestHash to where it is used
